Move integration event dispatcher interfaces to infrastructure layer
