+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Integration/Messages/MeteringPointCreatedEventMessage.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.MeteringPoints.Application.Common.Transport;
-using Energinet.DataHub.MeteringPoints.Application.IntegrationEvent;
+using Energinet.DataHub.MeteringPoints.Infrastructure.Integration.Dispatchers;
 using MediatR;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.Integration.Messages
